// frontend/src/services/api-client/models/story-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Story Generator
 * The story generator API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ChapterDto } from './chapter-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { PublishingDto } from './publishing-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { TypeDto } from './type-dto';

/**
 * 
 * @export
 * @interface StoryDto
 */
export interface StoryDto {
    /**
     * ID of the story
     * @type {number}
     * @memberof StoryDto
     */
    'id'?: number;
    /**
     * Title of the story
     * @type {string}
     * @memberof StoryDto
     */
    'name': string;
    /**
     * Synopsis of the story
     * @type {string}
     * @memberof StoryDto
     */
    'synopsis': string;
    /**
     * 
     * @type {string}
     * @memberof StoryDto
     */
    'audio_url': string;
    /**
     * 
     * @type {string}
     * @memberof StoryDto
     */
    'image_prompt': string;
    /**
     * 
     * @type {string}
     * @memberof StoryDto
     */
    'background_image': string;
    /**
     * 
     * @type {string}
     * @memberof StoryDto
     */
    'thumbnail_url': string;
    /**
     * 
     * @type {string}
     * @memberof StoryDto
     */
    'video_url': string;
    /**
     * Array of chapters in the story
     * @type {Array<ChapterDto>}
     * @memberof StoryDto
     */
    'chapters': Array<ChapterDto>;
    /**
     * 
     * @type {Array<PublishingDto>}
     * @memberof StoryDto
     */
    'publishings'?: Array<PublishingDto>;
    /**
     * ID of the story type
     * @type {number}
     * @memberof StoryDto
     */
    'types_id': number;
    /**
     * 
     * @type {TypeDto}
     * @memberof StoryDto
     */
    'types': TypeDto;
    /**
     * 
     * @type {string}
     * @memberof StoryDto
     */
    'created_at'?: string;
    /**
     * 
     * @type {string}
     * @memberof StoryDto
     */
    'updated_at'?: string;
}